refactor: extract TriageItemCycleWorkflow as singleton sub-workflow
Split per-item triage processing (context → panel → PO → labels)
into TriageItemCycleWorkflow which runs as a singleton. Elsa queues
dispatches so items are triaged sequentially, preventing LLM overload.

IssueTriageWorkflow now just fetches items and dispatches fire-and-forget
cycles. Backpressure propagates through the singleton constraint and
LlmCallWorkflow's self-throttling on budget/concurrency limits.

Update wiki: triage page, new triage-item-cycle page, Workflows.md
index (30 workflows), dependency diagram, React Flow diagrams.
